+++ apps/astarte_data_updater_plant/test/support/cases/device.ex
@@ -261,6 +261,10 @@ defmodule Astarte.DataUpdaterPlant.Cases.Device do
        end},
       {:fixed_endpoint_interface,
        fn acc -> [new_interfaces(fixed_endpoint_interface(), acc, :single)] end},
+      {:fixed_object_datastream_1,
+       fn acc -> [new_interfaces(fixed_object_datastream_1(), acc, :single)] end},
+      {:fixed_object_datastream_2,
+       fn acc -> [new_interfaces(fixed_object_datastream_2(), acc, :single)] end},
       {
         :encrypted_endpoints_properties_interfaces,
         fn acc ->
@@ -290,6 +294,8 @@ defmodule Astarte.DataUpdaterPlant.Cases.Device do
 
     %{
       interfaces: all_interfaces,
+      fixed_object_datastream_1: named_interfaces.fixed_object_datastream_1,
+      fixed_object_datastream_2: named_interfaces.fixed_object_datastream_2,
       fixed_endpoint_interface: named_interfaces.fixed_endpoint_interface,
       server_property_with_all_endpoint_types:
         named_interfaces.server_property_with_all_endpoint_types,
@@ -416,6 +422,58 @@ defmodule Astarte.DataUpdaterPlant.Cases.Device do
       aggregation: :object,
       mappings: mappings
     )
+  end
+
+  defp fixed_object_datastream_1 do
+    interface(
+      name: "test.FixedObjectDatastream1",
+      ownership: :device,
+      type: :datastream,
+      aggregation: :object
+    )
+    |> map(fn interface ->
+      mapping = Enum.at(interface.mappings, 0)
+
+      mapping_first = %{
+        mapping
+        | endpoint: "/data/first",
+          value_type: :string
+      }
+
+      mapping_second = %{
+        mapping
+        | endpoint: "/data/second",
+          value_type: :integer
+      }
+
+      %{interface | mappings: [mapping_first, mapping_second]}
+    end)
+  end
+
+  defp fixed_object_datastream_2 do
+    interface(
+      name: "test.FixedObjectDatastream2",
+      ownership: :device,
+      type: :datastream,
+      aggregation: :object
+    )
+    |> map(fn interface ->
+      mapping = Enum.at(interface.mappings, 0)
+
+      mapping_first = %{
+        mapping
+        | endpoint: "/data/first",
+          value_type: :boolean
+      }
+
+      mapping_second = %{
+        mapping
+        | endpoint: "/data/second",
+          value_type: :binaryblob
+      }
+
+      %{interface | mappings: [mapping_first, mapping_second]}
+    end)
   end
 
   defp all_endpoint_types(ownership, type) do